+++ src/services/__init__.py
@@ -1,5 +1,1 @@
 """Services package for game screen selection logic."""
-
-from .screen_picker import GameScreenPicker
-
-__all__ = ["GameScreenPicker"]
