+++ src/utils/__init__.py
@@ -1,6 +1,1 @@
 """Utils package for file operations and vector utilities."""
-
-from .file_utils import FileUtils
-from .vector_utils import VectorUtils
-
-__all__ = ["FileUtils", "VectorUtils"]
